runfix: start migration flow in background after app was initialised
