+++ src/script/main/app.ts
@@ -80,7 +80,6 @@ import {IntegrationService} from '../integration/IntegrationService';
 import {startNewVersionPolling} from '../lifecycle/newVersionHandler';
 import {MediaRepository} from '../media/MediaRepository';
 import {initMLSCallbacks, initMLSConversations, registerUninitializedSelfAndTeamConversations} from '../mls';
-import {initialiseMLSMigrationFlow} from '../mls/MLSMigration';
 import {NotificationRepository} from '../notification/NotificationRepository';
 import {PreferenceNotificationRepository} from '../notification/PreferenceNotificationRepository';
 import {PermissionRepository} from '../permission/PermissionRepository';
@@ -166,8 +165,8 @@ export class App {
    * @param apiClient Configured backend client
    */
   constructor(
-    private readonly core: Core,
-    private readonly apiClient: APIClient,
+    public readonly core: Core,
+    public readonly apiClient: APIClient,
     private readonly config: Configuration,
   ) {
     this.config = config;
@@ -437,20 +436,6 @@ export class App {
 
         //add the potential `self` and `team` conversations
         await registerUninitializedSelfAndTeamConversations(conversations, selfUser, clientEntity().id, this.core);
-
-        //FIXME: this value will be read from team feature config
-        const mockedMigrationConfig = {
-          clientsThreshold: 100,
-          usersThreshold: 100,
-          finaliseRegardlessAfter: 1683885139353,
-          startTime: 1683280357523,
-        };
-
-        await initialiseMLSMigrationFlow(mockedMigrationConfig, conversations, {
-          core: this.core,
-          apiClient: this.apiClient,
-          isConversationOwnedBySelfTeam: ({team_id}) => !!selfUser.teamId && team_id === selfUser.teamId,
-        });
       }
 
       telemetry.timeStep(AppInitTimingsStep.UPDATED_FROM_NOTIFICATIONS);
